Add Band controller, refactor routes
