+++ lib/routers/bands.js
@@ -1,8 +1,6 @@
 const express = require('express');
-const Band = require('../models/Band');
-const User = require('../models/User')
-const MembershipRequest = require('../models/MembershipRequest')
 const passport = require('passport');
+const BandController = require('../controllers/BandController')
 
 const router = express.Router();
 
@@ -12,28 +10,7 @@ router.post('/bands', async function(req, res, next) {
       return next(err)
     }
     if(user){
-      try {
-	var params = bandParamsFromReqBody(req.body, user)
-      } catch (e) {
-	return res.status(400).json("User profile incomplete")
-      }
-      let band = new Band(params)
-      band.save().then(()=>{
-        return User.findByIdAndUpdate(
-          user._id,
-          { $addToSet: { bands: band._id } },
-          { new: true, useFindAndModify: false },
-          (err, user) => {
-            if (err) {
-              return res.status(500).json("Update user object failed")
-            }
-            if (!user) {
-              return res.status(422).json("Find user failed")
-            }
-            return res.status(201).json("Band created successfully")
-          })
-      })
-      .catch((err)=>{res.status(500).send(err)})
+      BandController.createBandAction(user, req.body, res, next)
       return
     }
     return res.status(401).json(info)
@@ -46,44 +23,11 @@ router.get('/bands', async function(req, res, next) {
       return next(err)
     }
     if(user){
-      if (!user.location){
-	return res.status(400).json("User Profile Not Set")
-      }
-      let findParams = []
-      user.instruments.map((instrument)=>{
-        findParams.push({ [`missingInstruments.${instrument.instrument}`]: { $exists : true } })
-      })
-      let memberships = await MembershipRequest.find({user_id: user.id}, {band_id: 1, _id: 0})
-      memberships = memberships.map((band)=>{
-        return band.band_id
-      })
-      let bands = await Band.find()
-                            .or(findParams)
-                            .and({
-                              location: {
-                                $near: {
-                                  $geometry: {
-                                    type: "Point",
-                                    coordinates: user.location.coords
-                                  },
-                                  $maxDistance: 2000
-                                },
-                              }
-                            }).and ({
-                              _id:{$nin:memberships}
-                            })
-      return res.status(200).send({bands: bands})
+      BandController.viewBandsAction(user, res)
+      return
     }
     return res.status(401).json(info)
   })(req, res, next)
 })
-
-const bandParamsFromReqBody = (body, user) => {
-  body.location = {}
-  body.location.coordinates = user.location.coords
-  body.locationFriendly = user.location.friendlyLocation
-  body.user = user.id
-  return body
-}
 
 module.exports = router;
